+++ changes.md
@@ -1,2 +1,3 @@
-* leaflet-elevation 2.2.6
-* gesture-handling language
+* Solved Bug: detect Browser language in gesture handling
+* leaflet-elevation-2.2.6
+* leaflet-gesture-handling: leaflet-control-* (code opt.)