+++ changes.md
@@ -1,1 +1,2 @@
 * leaflet-gestures-1.4.3
+* Leaflet.Control.FullScreen 2.4.0